[feat] UserInGame 분리
